기록 업데이트

+++ WE/Sources/GameFramework/GameCore.cpp
@@ -94,7 +94,7 @@ namespace GameCore
 
 				GetInputSystemManager()->Tick();
 
-				Game.Update(Timer->GetDeltaTime());
+				Game.Update((float)Timer->GetDeltaTime());
 			}
 		}
 	}
