+++ WE/Sources/GameFramework/Object/Pawn/Pawn.cpp
@@ -7,7 +7,7 @@ void APawn::Tick_PrePhysics(float Delta)
 
 void APawn::AddMovementInput(const XMFLOAT3& WorldDirection, float ScaleValue)
 {
-	ScaleValue *= GetEngineTimer()->GetDeltaTime() * mMoveSpeed;
+	ScaleValue *= (float)GetEngineTimer()->GetDeltaTime() * mMoveSpeed;
 	XMFLOAT3 WorldLocation = GetActorTransform().Translation;
 	XMVECTOR DirectionVector = XMLoadFloat3(&WorldDirection);
 	DirectionVector = DirectX::XMVectorScale(DirectionVector, ScaleValue);
